fix cast warning

+++ AudioMirror/MiniportWaveRTStream.cpp
@@ -405,7 +405,7 @@ NTSTATUS MiniportWaveRTStream::AllocateBufferWithNotification
 	m_ulNotificationIntervalMs = ulBufferDurationMs / NotificationCount_;
 
 	m_RingBuffer = new(NonPagedPoolNx, MINWAVERTSTREAM_POOLTAG)RingBuffer;
-	m_RingBuffer->Init(m_ulDmaBufferSize * 4, m_pWfExt->Format.nBlockAlign);
+	m_RingBuffer->Init((SIZE_T)m_ulDmaBufferSize * 4, m_pWfExt->Format.nBlockAlign);
 
 	*AudioBufferMdl_ = pBufferMdl;
 	*ActualSize_ = RequestedSize_;
